Added large file support to the plugin
- Large files (>2GB) can now be imported - as their data is now stored in TArray64.
- When importing large files, we no longer store the raw data for potential export later on.
  (this is to avoid creating unnecessary large uasset files)
- Converted internal code to use TArray64 when necessary, as well as uint64/int64 when needed.
- Unfortunately, it seems that Niagara still mainly uses 32bits indexing internally,
  or for shader generation and parameters, so we're still limited to a 32bits max size for
  combined number of points - frames - exported attributes.
- The new serialization code for large file is backward compatible with previous version of the point caches assets.
- Optimized raw data compression for small files.
- Fixed warnings in the GPU code due to the Alpha/Impulse/Type function incorrectly using float3 instead of floats.
- If the amount of data imported in the plugin's buffers is too large too be processed afterwards
  by Niagara (due to 32bits buffer / indexing) output an error in the log and fail the file import.

+++ Source/HoudiniNiagara/Public/HoudiniPointCacheLoaderJSONBase.h
@@ -24,6 +24,8 @@
 #pragma once
 
 #include "CoreMinimal.h"
+#include "RHIDefinitions.h"
+
 #include "HoudiniPointCacheLoader.h"
 
 class UHoudiniPointCache;
@@ -31,9 +33,9 @@ class UHoudiniPointCache;
 struct FHoudiniPointCacheJSONHeader 
 {
     FString Version;
-    uint32 NumSamples;
+    uint64 NumSamples;
     uint32 NumFrames;
-    uint32 NumPoints;
+    uint64 NumPoints;
     uint32 NumAttributes;
     uint32 NumAttributeComponents;
     TArray<FString> Attributes;
